Rework IQL grammar and the gradle setup.
For testing purposes we now split the grammar into the public base
grammar 'IQL' and an inherited grammar 'IQL_Test' which holds all the
rules to be tested in a form that forces them to match the entire test
input. This way we can test those rules in as if they were embedded into
a greater context.