fix(model): add direct manifest linking to StaticStructure.
StaticStructure was using an old approach to keeping link count down and
fetch the manifest for a structure. Due to the way this method was
implemented it relied heavily on a very specific composition of
manifests and created issues for test setups. We now store the link in
every structure object directly.